Add files via upload
The object of hangman is to guess the secret word before the stick figure is hung. 
Players take turns selecting letters to narrow the word down. 
Players can take turns or work together. 
Gameplay continues until the players guess the word or they run out of guesses and the stick figure is hung.